backend/requirements.txt
# SmartTag back-end / ML smoke (Python 3.11+)
numpy>=1.26
scikit-learn>=1.4
joblib>=1.3
paho-mqtt>=2.0
psycopg[binary]>=3.1
python-dotenv>=1.0
fastapi>=0.115
uvicorn[standard]>=0.30
